Remove upgrading of packages from backports repositories
It could cause situations when some packages e.g systemd-container (v.232)
cannot be installed due to version mismatch with package which it based on
e.g. systemd (v.234). In this case it looks like that package
systemd-container was not backported yet and cannot be installed on top
of newer systemd